Finish decode Data Processing Register and tests
Some test fixes in encode branches

+++ decode/buildersDecode/decodeDPI/buildInstructionDPI.c
@@ -10,7 +10,7 @@ Instruction buildADDI(uint32_t data) {
     instruction.type = (uint8_t)GET_BITS(data, 0xF, 1, 25);
     instruction.rd = (uint8_t)GET_BITS(data, 0x1F, 0, 0);
     instruction.rn = (uint8_t)GET_BITS(data, 0x1F, 1, 5);
-    instruction.imm = (uint8_t)GET_BITS(data, 0xFFF, 1, 10);
+    instruction.imm = (uint16_t)GET_BITS(data, 0xFFF, 1, 10);
 
     return instruction;
 }
@@ -20,7 +20,7 @@ Instruction buildSUBI(uint32_t data) {
     instruction.type = (uint8_t)GET_BITS(data, 0xF, 1, 25);
     instruction.rd = (uint8_t)GET_BITS(data, 0x1F, 0, 0);
     instruction.rn = (uint8_t)GET_BITS(data, 0x1F, 1, 5);
-    instruction.imm = (uint8_t)GET_BITS(data, 0xFFF, 1, 10);
+    instruction.imm = (uint16_t)GET_BITS(data, 0xFFF, 1, 10);
 
     return instruction;
 }
@@ -30,7 +30,7 @@ Instruction buildCMP(uint32_t data) {
     instruction.type = (uint8_t)GET_BITS(data, 0xF, 1, 25);
     instruction.rn = (uint8_t)GET_BITS(data, 0x1F, 1, 5);
     instruction.rd = (uint8_t)GET_BITS(data, 0x1F, 0, 0);
-    instruction.imm = (uint8_t)GET_BITS(data, 0xFFF, 1, 10);
+    instruction.imm = (uint16_t)GET_BITS(data, 0xFFF, 1, 10);
 
     return instruction;
 }
